pyproject.toml
[tool.poetry]
name = "birthdaycongratsgenerator"
version = "0.1.0"
description = ""
authors = ["Alexander Lubovsky <alexander.lubovsky@intexsoft.by>"]

[tool.poetry.dependencies]
python = "3.10.5"
python-telegram-bot = "^13.12"
python-dotenv = "^0.20.0"
flake8 = "^4.0.1"
black = "^22.3.0"
isort = "^5.10.1"
pre-commit = "^2.19.0"
autoflake = "^1.4"
pyright = "^1.1.255"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
